Fix dbunit tests
The adhoc migrations in the dbunit tests rely on implicit constraint
drop cascade when dropping the message table. Adding an explicit
constraint drop makes the tests a bit more stable. Separating the
individual statements into separate try-catch blocks further adds to the
stability of the tests.

The database name in the dbunit tests was still referring to the tests
copy-paste parent, the liquibase module tests. Switching to "dbunit" for
consistency.